Refactor colourisation of errors, add more cases to highlight
Use variables to contain the regular expressions to match, this eases future extensions.

In particular, add support for latex error messages preceded by !, as in
    ! Undefined control sequence.
    l.582 \foo
or by preceded with line numbers, using latex's `-file-line-error` option, as in
    ./01_introduction.tex:582: Undefined control sequence.
    l.582 \foo

Adding the `[ $COLOR -eq 1]` test prevents useless RegExp match and replace when nothing is changed.